⚠️ first pass table data
