+++ dashboard-app/src/App.tsx
@@ -5,7 +5,7 @@ import { Route, Switch, BrowserRouter, withRouter } from 'react-router-dom';
 import PrivateRoute from './auth/components/PrivateRoute';
 import HoldingPage from './HoldingPage';
 import ByActivity from './dashboard/ByActivity';
-import ByTime from './dashboard/ByTime';
+import ByTime from './dashboard/ByTime/index';
 import ByVolunteer from './dashboard/ByVolunteer';
 import Dashboard from './dashboard/Dashboard';
 import Login from './auth/pages/Login';
